requirements.txt
etl-entities>=1.0,<1.1
ftputil
hdfs[kerberos]>2
humanize
mts_apache_atlas>=0.3.3,<0.4
paramiko
platformdirs
pydantic
PySmbClient
pyyaml
typing_extensions
